Split cookie protections
