+++ shared/js/content-scope/fingerprintingAudioProtection.js
