+++ shared/js/content-scope/fingerprintingBatteryProtection.js
