+++ shared/js/content-scope/fingerprintingCanvasProtection.js
