+++ shared/js/content-scope/fingerprintingHardwareProtection.js
