+++ shared/js/content-scope/fingerprintingScreenSizeProtection.js
