+++ shared/js/content-scope/fingerprintingTemporaryStorageProtection.js
