+++ shared/js/content-scope/gpcProtection.js
