+++ shared/js/content-scope/referrerProtection.js
